+++ src/logger.ts
@@ -6,6 +6,7 @@ enum LogLevel {
   INFO = 'info',
   WARNING = 'warning',
   ERROR = 'error',
+  NONE = 'none',
 }
 
 let LOGGER: Logger;
@@ -13,10 +14,22 @@ let LOGGER: Logger;
 class Logger implements vscode.Disposable {
   private outputChannel: vscode.OutputChannel;
   private level: LogLevel;
+  private onDidChangeConfigurationDisposable: vscode.Disposable;
 
-  constructor(channelName: string, level: LogLevel = LogLevel.INFO) {
+  constructor(channelName: string) {
     this.outputChannel = vscode.window.createOutputChannel(channelName);
-    this.level = level;
+    this.level =
+      vscode.workspace.getConfiguration(EXTENSION_NAME).get(SETTINGS.logLevel) ??
+      LogLevel.INFO;
+    this.onDidChangeConfigurationDisposable = vscode.workspace.onDidChangeConfiguration(
+      (e) => {
+        if (e.affectsConfiguration(`${EXTENSION_NAME}.${SETTINGS.logLevel}`)) {
+          this.level =
+            vscode.workspace.getConfiguration(EXTENSION_NAME).get(SETTINGS.logLevel) ??
+            this.level;
+        }
+      },
+    );
   }
 
   public info(message: string) {
@@ -37,13 +50,19 @@ class Logger implements vscode.Disposable {
 
   public dispose() {
     this.outputChannel.dispose();
+    this.onDidChangeConfigurationDisposable.dispose();
   }
 
   private log(message: string, level: LogLevel = LogLevel.INFO) {
     if (!this.shouldLog(level)) return;
 
     const timestamp = new Date().toISOString();
-    this.outputChannel.append(`[${timestamp}] [${level}] ${message}`);
+    const logMessage = `[${timestamp}] [${level}] ${message}`;
+    if (message.endsWith('\n')) {
+      this.outputChannel.append(logMessage);
+    } else {
+      this.outputChannel.appendLine(logMessage);
+    }
   }
 
   private shouldLog(level: LogLevel): boolean {
@@ -53,11 +72,7 @@ class Logger implements vscode.Disposable {
 
 export const getLogger = (): Logger => {
   if (!LOGGER) {
-    LOGGER = new Logger(
-      EXTENSION_NAME,
-      vscode.workspace.getConfiguration(EXTENSION_NAME).get(SETTINGS.logLevel) ??
-        LogLevel.INFO,
-    );
+    LOGGER = new Logger(EXTENSION_NAME);
   }
   return LOGGER;
 };
@@ -70,6 +85,8 @@ const getLogLevelValue = (level: string): number => {
       return 1;
     case LogLevel.ERROR:
       return 2;
+    case LogLevel.NONE:
+      return 3;
     default:
       return 0;
   }